final-pjt-back/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
bardapi==0.1.11
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.18
django-allauth==0.52.0
django-cors-headers==4.0.0
django-extensions==3.2.1
djangorestframework==3.14.0
executing==1.2.0
fonttools==4.39.4
idna==3.4
importlib-resources==5.12.0
ipython==8.13.2
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
threadpoolctl==3.1.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
zipp==3.15.0
